LF-3747 extract getComparator from Table file
