+++ packages/webapp/src/components/Table/v2/index.jsx
@@ -26,6 +26,7 @@ import { BsThreeDots } from 'react-icons/bs';
 import clsx from 'clsx';
 import EnhancedTableHead from './TableHead';
 import Button from '../../Form/Button';
+import { getComparator } from '../../../util/sort';
 import styles from './styles.module.scss';
 
 const More = ({ onClickLoadMore, onClickMore, invisibleRowCount, dense, colSpan }) => {
@@ -71,22 +72,6 @@ More.propTypes = {
   dense: PropTypes.bool,
   colSpan: PropTypes.number,
 };
-
-function descendingComparator(a, b, orderBy) {
-  if (b[orderBy] < a[orderBy]) {
-    return -1;
-  }
-  if (b[orderBy] > a[orderBy]) {
-    return 1;
-  }
-  return 0;
-}
-
-function getComparator(order, orderBy) {
-  return order === 'desc'
-    ? (a, b) => descendingComparator(a, b, orderBy)
-    : (a, b) => -descendingComparator(a, b, orderBy);
-}
 
 /**
  * A table component built utilizing the Material Ui Table.
